+++ src/njoy21/input/PLOTR/Card10a/Xpoint/test/Xpoint.test.cpp
@@ -11,7 +11,7 @@ using namespace njoy::njoy21::input;
 SCENARIO( "Xpoint output values", "[PLOTR],[Card10a], [Xpoint]"){
 
   GIVEN( "valid Xpoint parameters" ){
-    std::vector<float> validValues{-10.0,-0.3,0.0,0.1,0.5,0.8,1.0,2.0,80.0};
+    std::vector<double> validValues{-10.0,-0.3,0.0,0.1,0.5,0.8,1.0,2.0,80.0};
 
     THEN( "the returned class has the correct value" ){
       for( auto xpoint : validValues ){
