+++ src/njoy21/input/PLOTR/Card4/Ytag/test/Ytag.test.cpp
@@ -11,7 +11,7 @@ using namespace njoy::njoy21::input;
 SCENARIO( "Ytag output values", "[PLOTR],[Card4], [Ytag]"){
 
   GIVEN( "valid Ytag parameters" ){
-    std::vector<float> validValues{-2.0,-1.1,0.0,0.1,0.5,0.8,1.0,2.0,80.0};
+    std::vector<double> validValues{-2.0,-1.1,0.0,0.1,0.5,0.8,1.0,2.0,80.0};
 
     THEN( "the returned class has the correct value" ){
       for( auto ytag : validValues ){
