+++ src/njoy21/input/PLOTR/Card7/Rtop.hpp
@@ -1,5 +1,5 @@
 struct Rtop {
-  using Value_t = float;
+  using Value_t = double;
   static std::string name(){ return "rtop"; }
   static std::string description(){
     return
@@ -7,11 +7,11 @@ struct Rtop {
       "or the z-axis.\n\n"
       "Checks that rtop has a greater value than rbot.";
   }
-  static Value_t defaultValue( const float rbot ){
+  static Value_t defaultValue( const double rbot ){
     return rbot;
   }
   static bool verify( const Value_t v, 
-                      const float rbot ){
+                      const double rbot ){
     return ( v >= rbot );
   }
 };
